+++ R/evaluate_effects.R
@@ -6,6 +6,8 @@
 #' @param iv An independent variable in the model.
 #' @param dv A dependent variable in the model.
 #'
+#' @return A matrix of the estimated F Square metric for each construct.
+#'
 #' @usage
 #' fSquared(seminr_model, iv, dv)
 #'
@@ -57,10 +59,15 @@ fSquared <- function(seminr_model, iv, dv) {
   #   lm <- stats::lm(formula = frmla, data = data)
   #   summary(lm)
   #   }
-  utils::capture.output(
+  suppressMessages(
     without_pls <- estimate_pls(data = seminr_model$rawdata,
                                 measurement_model = seminr_model$measurement_model,
-                                structural_model = without_sm)
+                                structural_model = without_sm,
+                                inner_weights = seminr_model$inner_weights,
+                                missing = seminr_model$settings$missing,
+                                missing_value = seminr_model$settings$missing_value,
+                                maxIt = seminr_model$settings$maxIt,
+                                stopCriterion = seminr_model$settings$stopCriterion)
   )
   with_r2 <- seminr_model$rSquared["Rsq", dv]
   ifelse(any(without_sm[,"target"] == dv),
@@ -71,21 +78,22 @@ fSquared <- function(seminr_model, iv, dv) {
 }
 
 model_fsquares <- function(seminr_model) {
-  if (any(names(seminr_model$measurement_model) == "orthogonal_interaction")
-      | any(names(seminr_model$measurement_model) == "two_stage_interaction")
-      | any(names(seminr_model$measurement_model) == "scaled_interaction" )) {
-    return("The fSquare cannot be calculated as the model contains an interaction term and omitting either the antecedent or moderator in the interaction term will cause model estimation to fail")
-  }
-
   path_matrix <- seminr_model$path_coef
   fsquared_matrix <- path_matrix
-
   for (dv in all_endogenous(seminr_model$smMatrix)) {
-    for (iv in all_exogenous(seminr_model$smMatrix)) {
+    ifelse(length(interactions_of(dv, seminr_model$smMatrix) ) > 0,
+      int_components <- unique(unlist(strsplit(interactions_of(dv, seminr_model$smMatrix), "\\*"))),
+      int_components <- NA)
+    for (iv in setdiff(all_exogenous(seminr_model$smMatrix), int_components)) {
       fsquared_matrix[iv, dv] <- fSquared(seminr_model = seminr_model,
                                           iv = iv,
                                           dv = dv)
+      fsquared_matrix[int_components,dv] <- NA
     }
   }
+  if (length(all_interactions(seminr_model$smMatrix) > 0)) {
+    comment(fsquared_matrix) <- "The fSquare for certain relationships cannot be calculated as the model contains an interaction term and omitting either the antecedent or moderator in the interaction term will cause model estimation to fail"
+  }
   convert_to_table_output(fsquared_matrix)
+
 }